+++ src/nos/utils/__init__.py
@@ -1,5 +1,8 @@
 from .nested_dataclass_to_dict import (
     dataclass_to_dict,
 )
+from .unique_id import (
+    UniqueId,
+)
 
-__all__ = ["dataclass_to_dict"]
+__all__ = ["dataclass_to_dict", "UniqueId"]
